Allow IInstructService to StopCompletion.

+++ src/Plugins/BotSharp.Plugin.LLamaSharp/LlamaAiModel.cs
@@ -35,7 +35,7 @@ public class LlamaAiModel
 
         _params = new ModelParams(Path.Combine(_settings.ModelDir, model))
         {
-            ContextSize = _settings.MaxContextLength,
+            ContextSize = (uint)_settings.MaxContextLength,
             Seed = 1337,
             GpuLayerCount = _settings.NumberOfGpuLayer
         };
